projects/phase-06-research-assistant/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
openai>=1.0.0
python-dotenv>=1.0.0
rich>=13.0.0
duckduckgo-search>=4.0.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
markdownify>=0.11.0
